pyproject.toml
[project]
name = "tulip-api"
version = "0.1.0"
description = "FastAPI service for TULIP model embeddings"
authors = []
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "torch>=2.0.0",
    "timm>=0.9.0",
    "transformers>=4.35.0",
    "Pillow>=10.0.0",
    "python-multipart>=0.0.6",  # Required for FastAPI file uploads
    "numpy>=1.24.0",
    "open-clip-torch @ git+https://github.com/tulip-berkeley/open_clip.git@main",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",  # For async API testing
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
include = [
    "src/tulip_api/**/*.py",
    "src/tulip_api/**/*.pyi",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tulip_api"]

[tool.ruff]
line-length = 100
target-version = "py38"
src = ["src"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto" 